# docs/swagger.yaml
basePath: /
definitions:
  api.CreateTemplateRequest:
    properties:
      body:
        type: string
      channel:
        $ref: '#/definitions/domain.Channel'
      metadata:
        additionalProperties: true
        type: object
      name:
        type: string
      subject:
        type: string
    type: object
  api.ErrorResponse:
    properties:
      details:
        example: '''to'' field is required'
        type: string
      error:
        example: validation failed
        type: string
    type: object
...
  api.SendNotificationRequest:
    properties:
      body:
        type: string
      channel:
        $ref: '#/definitions/domain.Channel'
      data:
        additionalProperties: true
        type: object
      idempotency_key:
        type: string
      priority:
        $ref: '#/definitions/domain.Priority'
      recipient_id:
        type: string
      recipient_name:
        type: string
      scheduled_for:
        type: string
      subject:
        type: string
      template:
        type: string
      to:
        type: string
    type: object
  api.UpdateTemplateRequest:
    properties:
      active:
        type: boolean
      body:
        type: string
      metadata:
        additionalProperties: true
        type: object
      subject:
        type: string
    type: object
  domain.Channel:
    enum:
    - email
    - sms
    - push
    - in_app
    type: string
    x-enum-varnames:
    - ChannelEmail
    - ChannelSMS
    - ChannelPush
    - ChannelInApp
  domain.JSONB:
    additionalProperties: true
    type: object
  domain.Notification:
    properties:
      batch_id:
        type: string
      body:
        type: string
      channel:
        $ref: '#/definitions/domain.Channel'
      created_at:
        type: string
      created_by:
        type: string
      deleted_at:
        type: string
      delivered_at:
        type: string
      error_code:
        type: string
      error_message:
        type: string
      failed_at:
        type: string
      html_body:
        type: string
      id:
        type: string
      max_retries:
        type: integer
      metadata:
        $ref: '#/definitions/domain.JSONB'
      next_retry_at:
        type: string
      priority:
        $ref: '#/definitions/domain.Priority'
      provider:
        type: string
      provider_message_id:
        type: string
      provider_response:
        $ref: '#/definitions/domain.JSONB'
      queued_at:
        type: string
      recipient:
        $ref: '#/definitions/domain.JSONB'
      retry_count:
        type: integer
      sent_at:
        type: string
      status:
        $ref: '#/definitions/domain.NotificationStatus'
      subject:
        type: string
      template_id:
        type: string
      updated_at:
        type: string
      variables:
        $ref: '#/definitions/domain.JSONB'
    type: object
  domain.NotificationStatus:
    enum:
    - pending
    - queued
    - processing
    - sent
    - delivered
    - failed
    - cancelled
    type: string
    x-enum-varnames:
    - StatusPending
    - StatusQueued
    - StatusProcessing
    - StatusSent
    - StatusDelivered
    - StatusFailed
    - StatusCancelled
  domain.Priority:
    enum:
    - low
    - normal
    - high
    - urgent
    type: string
    x-enum-varnames:
    - PriorityLow
    - PriorityNormal
    - PriorityHigh
    - PriorityUrgent
  domain.Template:
    properties:
      body:
        type: string
      channels:
        items:
          type: string
        type: array
      config:
        $ref: '#/definitions/domain.JSONB'
      created_at:
        type: string
      created_by:
        type: string
      default_values:
        $ref: '#/definitions/domain.JSONB'
      deleted_at:
        type: string
      description:
        type: string
      html_body:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      name:
        type: string
      subject:
        type: string
      updated_at:
        type: string
      variables:
        items:
          type: string
        type: array
    type: object
host: localhost:8080
info:
  contact:
    email: support@yourdomain.com
    name: API Support
  description: Unified notification delivery service supporting email, SMS, push notifications,
    and in-app messaging with bulk sending capabilities.
  title: Buzz Notification Service API
  version: 1.0.0
paths:
...
  /api/v1/notifications:
    get:
      description: Retrieve a paginated list of notifications with optional filters
      parameters:
      - description: Filter by status (queued, sent, delivered, failed)
        in: query
        name: status
        type: string
      - description: Filter by channel (email, sms, push, in_app)
        in: query
        name: channel
        type: string
      - description: Filter by recipient ID
        in: query
        name: recipient_id
        type: string
      - description: Page size 1-100 (default 20)
        in: query
        name: limit
        type: integer
      - description: Page offset (default 0)
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - Bearer: []
      summary: List notifications
      tags:
      - notifications
    post:
      consumes:
      - application/json
      description: Queue a notification for delivery via the specified channel (email,
        sms, push, in_app)
      parameters:
      - description: Notification payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.SendNotificationRequest'
      produces:
      - application/json
      responses:
        "202":
          description: Notification queued
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Template not found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - Bearer: []
      summary: Send a notification
      tags:
      - notifications
  /api/v1/notifications/{id}:
    get:
      description: Retrieve a single notification by its UUID
      parameters:
      - description: Notification UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Notification'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - Bearer: []
      summary: Get a notification
      tags:
      - notifications
  /api/v1/templates:
    get:
      description: Retrieve a paginated list of notification templates with optional
        filters
      parameters:
      - description: Filter by channel (email, sms, push, in_app)
        in: query
        name: channel
        type: string
      - description: Filter by active status (default true)
        in: query
        name: active
        type: boolean
      - description: Page size (default 50)
        in: query
        name: limit
        type: integer
      - description: Page offset (default 0)
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - Bearer: []
      summary: List templates
      tags:
      - templates
    post:
      consumes:
      - application/json
      description: Create a reusable notification template with optional variable
        placeholders
      parameters:
      - description: Template definition
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.CreateTemplateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.Template'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - Bearer: []
      summary: Create a template
      tags:
      - templates
  /api/v1/templates/{name}:
    get:
      description: Retrieve a template by name
      parameters:
      - description: Template name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Template'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - Bearer: []
      summary: Get a template
      tags:
      - templates
    patch:
      consumes:
      - application/json
      description: Partially update an existing template's subject, body, metadata,
        or active status
      parameters:
      - description: Template name
        in: path
        name: name
        required: true
        type: string
      - description: Fields to update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.UpdateTemplateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - Bearer: []
      summary: Update a template
      tags:
      - templates
...
securityDefinitions:
  Bearer:
    description: API key prefixed with "Bearer ", e.g. "Bearer YOUR_API_KEY"
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"